backend/requirements.txt
# FastAPI Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.9  # CVE-2024-24762 (ReDoS) — was 0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic
pydantic>=2.10.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.11.0  # Required by plaid_service
passlib[argon2]==1.7.4
bcrypt==4.1.2
cryptography>=44.0.0  # Multiple CVEs in 42.x — was 42.0.0

# Background Jobs
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Financial APIs
plaid-python==17.0.0
yfinance==0.2.36  # Yahoo Finance for investment prices (FREE)
finnhub-python>=2.4.20  # Finnhub market data (free: 60 calls/min)
alpha_vantage>=3.0.0  # Alpha Vantage market data (free: 25 calls/day)

# Utilities
python-dateutil==2.8.2
pytz==2023.3
httpx==0.26.0

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7
sentry-sdk[fastapi]>=2.0.0  # v1 EOL — was 1.40.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
aiosqlite==0.20.0  # For in-memory SQLite testing

# Development
black==24.1.1
ruff==0.1.13
mypy==1.8.0

# Additional Testing & CI/CD
isort==5.13.2
flake8==7.0.0
pylint==3.0.3
safety==3.2.11  # Updated for pydantic 2.x compatibility
pre-commit==3.6.0

# Email
aiosmtplib>=3.0.0

# Multi-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2

# Cloud Storage (S3)
boto3>=1.34.0
